new metrics + degree distribution

+++ analysis.py
@@ -1,17 +1,19 @@
 import itertools
 import pickle
 import sys
-
+import json
 sys.path.append("graph_evolution")
-
+import networkx as nx
 from graph_evolution.organism import Organism
 # from organism import Organism
 from random import random,uniform
 import numpy as np
 from PIL import Image
-from utils_gan_flow import get_exp_measures
+from utils_gan_flow import get_exp_measures, plot_distributions
 import random
 import os
+
+from reference_metrics import get_degree_metric
 
 def array_to_greyscale_image(array: np.ndarray, output_path: str):
     """
@@ -38,8 +40,6 @@ def array_to_greyscale_image(array: np.ndarray, output_path: str):
     image.save(output_path)
 
 def compute_metrics(metric_names, population, dataset, run_name):
-
-
     fake_set = [organism.adjacencyMatrix for organism in population]
     test_set_path = f"GAN-flow/{dataset}/v_test.txt"
     with open(test_set_path, 'rb') as file:
@@ -51,31 +51,54 @@ def compute_metrics(metric_names, population, dataset, run_name):
     mixed_set_pairs = [pair for pair in itertools.product(uno , due)]
     len(fake_set), len(test_set), len(mixed_set_pairs), number_of_items
 
+    
+    os.makedirs(os.path.join(run_name, 'evaluations'), exist_ok=True)
     for metric_name in metric_names:
-
         # Load the already compute metrics
         test_metric_path = f"GAN-flow/{dataset}/experiments/{metric_name}/MoGAN/1.txt"
         with open(test_metric_path, 'rb') as file:
             test_metric = pickle.load(file)
 
-        fake_metric = get_exp_measures(fake_set, method=metric_name)
-        fake_metric_path = f"{run_name}/experiments/{metric_name}_fake.pkl"
-        with open(fake_metric_path, 'wb') as file:
-            pickle.dump(fake_metric, file)
+        fake_metric_path = f"{run_name}/evaluations/{metric_name}_fake.pkl"
+        if not os.path.exists(fake_metric_path):
+            fake_metric = get_exp_measures(fake_set, method=metric_name)
+            with open(fake_metric_path, 'wb') as file:
+                pickle.dump(fake_metric, file)
+        else:
+            with open(fake_metric_path, 'rb') as file:
+                fake_metric = pickle.load(file)
 
-        mixed_metric = get_exp_measures(mixed_set_pairs, paired = True,  method=metric_name)
-        mixed_metric_path = f"{run_name}/experiments/{metric_name}_mixed.pkl"
-        with open(mixed_metric_path, 'wb') as file:
-            pickle.dump(mixed_metric, file)
+        mixed_metric_path = f"{run_name}/evaluations/{metric_name}_mixed.pkl"
+        if not os.path.exists(mixed_metric_path):
+            mixed_metric = get_exp_measures(mixed_set_pairs, paired = True,  method=metric_name)
+            with open(mixed_metric_path, 'wb') as file:
+                pickle.dump(mixed_metric, file)
+        else:
+            with open(mixed_metric_path, 'rb') as file:
+                mixed_metric = pickle.load(file)
 
         mogan_metric_path = f"GAN-flow/{dataset}/experiments/{metric_name}/MoGAN/2.txt"
         with open(mogan_metric_path, 'rb') as file:
             mogan_metric = pickle.load(file)
 
-        plot_metrics(test_metric,fake_metric,mixed_metric, mogan_metric, metric_name, dataset)
+        plot_metrics(test_metric,fake_metric,mixed_metric, mogan_metric, metric_name, dataset, run_name)
 
+def get_distributions(distribution_names, population, dataset, run_name):
+    fake_set = [organism.adjacencyMatrix for organism in population]
+    test_set_path = f"GAN-flow/{dataset}/v_test.txt"
+    with open(test_set_path, 'rb') as file:
+        test_set = pickle.load(file)
 
-def plot_metrics(test_metric,fake_metric,mixed_metric, mogan_metric, metric_name, dataset):
+    if 'degree' in distribution_names:
+        test_dist, _ = get_degree_metric(test_set)
+        fake_dist, _ = get_degree_metric(fake_set)
+
+        plot_distributions([test_dist, fake_dist],
+                           ['Real', 'Fake'], 
+                           'Degree Distribution', 'Degree', 'Frequency', 
+                           f"{run_name}/evaluations/degree_distribution.png")
+
+def plot_metrics(test_metric,fake_metric,mixed_metric, mogan_metric, metric_name, dataset, run_name):
     import matplotlib.pyplot as plt
     plt.figure(figsize=(8,6))
     plt.hist(test_metric, bins = 50, alpha = 0.5, label = 'Real')
@@ -84,7 +107,7 @@ def plot_metrics(test_metric,fake_metric,mixed_metric, mogan_metric, metric_name
     plt.hist(mogan_metric, bins = 50, alpha = 0.5, label = 'MoGAN')
     plt.title(f'{metric_name} - {dataset}')
     plt.legend()
-    plt.savefig(f"results/{metric_name}_{dataset}.png")
+    plt.savefig(f"{run_name}/evaluations/{metric_name}_{dataset}.png")
 
 def get_all_npy_files(path):
     """
@@ -105,14 +128,24 @@ def get_all_npy_files(path):
 
 
 if __name__ == "__main__":
+    run_name = sys.argv[1]
     
-    run_name = 'data/sample'
-    with open(f'data/sample/0/final_pop.pkl', 'rb') as file:
-        obj_list = pickle.load(file)
+    if run_name[-1] == '/':
+        run_name.pop()
     
-    metrics_name= ['cutnorm']#['indegree', 'degree', 'topo', 'weight', 'cpc']
-    
-    compute_metrics(metrics_name, obj_list, 'BikeCHI', run_name)
+    with open(os.path.join(run_name,'config.json'), 'rb') as file:
+        config = json.load(file)
+
+    runs = config['reps']
+
+    for run in range(runs):
+        with open(os.path.join(run_name, str(run), 'final_pop.pkl'), 'rb') as file:
+            obj_list = pickle.load(file)
+        
+        metrics_name= ['indegree', 'degree', 'outdegree' ]#'topo', 'weight', 'cpc', 'cutnorm']
+        distribution_names = ['degree']
+        compute_metrics(metrics_name, obj_list, 'BikeCHI', os.path.join(run_name, str(run)))
+        get_distributions(distribution_names, obj_list, 'BikeCHI', os.path.join(run_name, str(run)))
 
 
     # for i,organism in enumerate(obj_list[:5]):
@@ -124,19 +157,19 @@ if __name__ == "__main__":
 
 
     # Example usage
-    npy_files = get_all_npy_files("GAN-flow/adj/BikeCHI")
+    # npy_files = get_all_npy_files("GAN-flow/adj/BikeCHI")
 
-    for i, npy_file in enumerate(npy_files[:10]):
-        npy_obj = np.load(npy_file)
-        npy_obj = npy_obj.astype(np.float64)
-        print("npy sum", np.sum(npy_obj))
-        npy_obj /= 327.0
+    # for i, npy_file in enumerate(npy_files[:10]):
+    #     npy_obj = np.load(npy_file)
+    #     npy_obj = npy_obj.astype(np.float64)
+    #     print("npy sum", np.sum(npy_obj))
+    #     npy_obj /= 327.0
 
-        array_to_greyscale_image(npy_obj, f"test/original{i}.png")
-        org = Organism(64, uniform(0, np.max(npy_obj)**3), [0, 327], npy_obj)
-        print("org sum", np.sum(org.adjacencyMatrix))
-        array_to_greyscale_image(org.adjacencyMatrix, f"test/rescaled{i}.png")
+    #     array_to_greyscale_image(npy_obj, f"test/original{i}.png")
+    #     org = Organism(64, uniform(0, np.max(npy_obj)**3), [0, 327], npy_obj)
+    #     print("org sum", np.sum(org.adjacencyMatrix))
+    #     array_to_greyscale_image(org.adjacencyMatrix, f"test/rescaled{i}.png")
 
-        diff = np.abs(npy_obj - np.array(org.adjacencyMatrix)/327)
-        print("diff sum", np.mean(diff)*327)
-        array_to_greyscale_image(diff/327, "test/diff.png")
+    #     diff = np.abs(npy_obj - np.array(org.adjacencyMatrix)/327)
+    #     print("diff sum", np.mean(diff)*327)
+    #     array_to_greyscale_image(diff/327, "test/diff.png")
